Adding main, constructor, and parse logic
This allows us to actually run the code that would use the private methods; for debugging purposes. SyntaxAnalyzer reads an input file (which is based on the output from the lexAnalyzer), then parse simply checks the PROG grammar rules and outputs the symbol-table if correct, or the current line if there is a syntax error.